pyproject.toml
[tool.poetry]
name = "trading-network-platform"
version = "0.1.0"
description = ""
authors = ["Pavel Akulich <pavelakulich1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg = {extras = ["binary"], version = "^3.2.1"}
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
django-filter = "^24.3"
redis = "^5.1.0"
celery = "^5.4.0"
qrcode = {extras = ["pil"], version = "^8.0"}
setuptools = "^75.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
